Work branch

+++ backend/src/main/java/com/jarom/funbankapp/controller/AccountController.java
@@ -211,7 +211,7 @@ public class AccountController {
     }
 
     // Reusable method to verify account ownership
-    private boolean userOwnsAccount(User user, int accountId) {
+    private boolean userOwnsAccount(User user, Long accountId) {
         System.out.println("🧪 Checking if user owns accountId: " + accountId);
         return accountDAO.findByUserId(user.getId())
                 .stream()
